+++ .gitignore
@@ -33,3 +33,5 @@ out/
 
 ### MAC ###
 .DS_Store
+
+intellij-java-google-style.xml
