+++ src/main/java/baseball/domain/BallPiece.java
@@ -1,0 +1,11 @@
+package baseball.domain;
+
+public class BallPiece {
+    final int number;
+    final int position;
+
+    public BallPiece(int number, int position) {
+        this.number = number;
+        this.position = position;
+    }
+}
