+++ src/main/java/baseball/domain/Hitter.java
@@ -1,0 +1,5 @@
+package baseball.domain;
+
+public interface Hitter {
+    BallContainer swing(Ball pitchBall);
+}
