+++ src/main/java/baseball/domain/HitterImpl.java
@@ -1,0 +1,16 @@
+package baseball.domain;
+
+import baseball.domain.service.BallGenerator;
+
+public class HitterImpl implements Hitter {
+    private final BallGenerator ballGenerator;
+
+    public HitterImpl(BallGenerator ballGenerator) {
+        this.ballGenerator = ballGenerator;
+    }
+
+    @Override
+    public BallContainer swing(Ball pitchBall) {
+        return new BallContainer(pitchBall, ballGenerator.generate());
+    }
+}
