+++ src/main/java/baseball/domain/Pitcher.java
@@ -1,0 +1,7 @@
+package baseball.domain;
+
+public interface Pitcher {
+    Ball pitch();
+
+    void setUpPitch();
+}
