--- src/main/java/baseball/domain/object/Hitter.java
@@ -1,8 +1,0 @@
-package baseball.domain.object;
-
-import baseball.domain.ball.Ball;
-import baseball.domain.ball.SwingBall;
-
-public interface Hitter {
-    SwingBall swing(Ball pitchBall);
-}
