--- src/main/java/baseball/domain/object/Pitcher.java
@@ -1,9 +1,0 @@
-package baseball.domain.object;
-
-import baseball.domain.ball.Ball;
-
-public interface Pitcher {
-    Ball pitch();
-
-    void setUpPitch();
-}
