+++ src/main/java/baseball/domain/service/BallGenerator.java
@@ -1,0 +1,10 @@
+package baseball.domain.service;
+
+import baseball.domain.Ball;
+
+public interface BallGenerator {
+    int MIN_NUMBER = 1;
+    int MAX_NUMBER = 9;
+
+    Ball generate();
+}
