--- src/main/java/baseball/enums/GameState.java
@@ -1,9 +1,0 @@
-package baseball.enums;
-
-public enum GameState {
-    PITCH_SETUP,
-    PITCH_WITH_SWING,
-    END_GAME,
-    SCORE_CALC,
-    RETRY_OR_EXIT
-}
