+++ src/main/java/baseball/game/enums/GameState.java
@@ -1,0 +1,9 @@
+package baseball.game.enums;
+
+public enum GameState {
+    PITCH_SETUP,
+    PITCH_WITH_SWING,
+    END_GAME,
+    SCORE_CALC,
+    RETRY_OR_EXIT
+}
