--- src/main/java/baseball/service/BallGenerator.java
@@ -1,10 +1,0 @@
-package baseball.service;
-
-import baseball.domain.ball.Ball;
-
-public interface BallGenerator {
-    int MIN_NUMBER = 1;
-    int MAX_NUMBER = 9;
-
-    Ball generate();
-}
